+++ dcf/revenue-exit_test.go
@@ -75,9 +75,9 @@ func TestRevenueExit(t *testing.T) {
 			},
 		}
 		currentDate := time.Date(2022, 5, 4, 0, 0, 0, 0, time.UTC)
-		revenueExitMultiple := float32(4.7)
+		revenueExitMultiple := float64(4.7)
 
-		var want float32 = 2186505.8
+		var want float64 = 2186505.8
 		got, _ := RevenueExit(&periods, currentDate, revenueExitMultiple)
 
 		if int(want/10) != int(got/10) {
@@ -88,7 +88,7 @@ func TestRevenueExit(t *testing.T) {
 	t.Run("Should return error if periods number are less than 2", func(t *testing.T) {
 		periods := []types.Period{}
 		currentDate := time.Date(2022, 5, 4, 0, 0, 0, 0, time.UTC)
-		revenueExitMultiple := float32(0)
+		revenueExitMultiple := float64(0)
 
 		_, err := RevenueExit(&periods, currentDate, revenueExitMultiple)
